requirements.txt
alembic==1.17.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==41.0.4
dnspython==2.8.0
email-validator==2.1.0
eventlet==0.33.3
flake8==7.3.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
greenlet==3.2.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
pycodestyle==2.14.0
pycparser==2.23
pydub==0.25.1
pyflakes==3.4.0
PyNaCl==1.5.0
pypng==0.20220715.0
python-dotenv==1.0.0
python-engineio==4.12.3
python-socketio==5.8.0
qrcode==7.4.2
requests==2.32.5
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.44
termcolor==3.1.0
typing_extensions==4.15.0
urllib3==2.5.0
Wave==0.0.2
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.2.1
